==> model/notebook/3 Geoparser/9 Address Normalization.ipynb <==
# Address Normalization
### By **Néstor Suat** in 2021

**Descripción:** Normalización de direcciones usando librería en C `libpostal`

Fuente: https://github.com/openvenues/libpostal

from postal.expand import expand_address
import pandas as pd
import re

## Variables para importar modelos y demás
dir_ = "../../data/v1/NER/"

file = 'tweets_location.tsv' # Dataset

## Importando Dataset
dataset = pd.read_csv(dir_+file, delimiter = "\t", quoting = 3)

locations = []
for i in range(len(dataset)):    
    loc = []
    text = dataset.iloc[i]['location']
    loc.append(text)
    expansions = expand_address(text, 
                                roman_numerals=False, 
                                split_alpha_from_numeric=True, 
                                expand_numex=False,
                                languages=["es"])
    
    expansions = re.sub(r'\b(\w+)( \1\b)+', r'\1', expansions[-1])
    loc = loc + [expansions]
    locations.append(loc)

df = pd.DataFrame(locations)
df.rename(columns={0:'original',1:'total'}, inplace=True)
df

df.to_csv(dir_+"tweets_location_libpostal_v3.tsv",sep='\t', index=False)

---

locations = []
for i in range(len(dataset)):    
    loc = []
    text = dataset.iloc[i]['location']
    loc.append(text)
    expansions = expand_address(text, 
                                roman_numerals=False, 
                                split_alpha_from_numeric=True, 
                                expand_numex=False,
                                languages=["es"])
    
    loc = loc + [len(expansions)] + expansions
    locations.append(loc)

from postal.expand import expand_address
t = "Av. 1ro de Mayo puente de la Av. 68"
res = expand_address(t, 
    roman_numerals=False, 
    split_alpha_from_numeric=True, 
    expand_numex=False,
    languages=["es"])[-1]
print(res)
re_output = re.sub(r'\b(\w+)( \1\b)+', r'\1', res)
print(re_output)

import re
sentence = 'Av. 1ro de Mayo puente de la Av. 68'

# remove duplicates
re_output = re.sub(r'\b(\w+)( \1\b)+', r'\1', sentence)
print('No duplicates:', re_output)
