==> model/notebook/3 Geoparser/11 Exploración.ipynb <==
# Exploración
### By **Néstor Suat** in 2021

**Descripción:** Explorando el conjunto de datos de **84.262 tweets de accidentes** con entidades reconocidas.

**Input:**
* accidents_tweets
---

import pandas as pd

#dir_ = '../../data/database/output_ml/M1/NER_extractor/'
dir_ = '../../data/database/output_ml/M1/NER_extractor/data_shielded_experiment/'

file_bogota = 'accident_tweets_unique_keywords'
dataset = pd.read_csv(dir_+file_bogota+'.tsv', delimiter = "\t", quoting = 3)
dataset.shape

dataset.head(5)

### Analizando ventanas de tiempo diferente
* Manualmente

# Octubre 2018

df = dataset[(dataset['created_at'] >= '2018-10-01') & (dataset['created_at'] < '2018-11-01')]
print("Min:", min(df['created_at']))
print("MAx:", max(df['created_at']))
print(df.shape)

### #Tweets por Dataset

df['dataset'].value_counts()

### #Tweets por username

df['user_name'].value_counts()

# Noviembre 2018

df = dataset[(dataset['created_at'] >= '2018-11-01') & (dataset['created_at'] < '2018-12-01')]
print("Min:", min(df['created_at']))
print("MAx:", max(df['created_at']))
print(df.shape)

### Tweets por Dataset
df['dataset'].value_counts()

### #Tweets por username
df['user_name'].value_counts()

# Diciembre 2018

df = dataset[(dataset['created_at'] >= '2018-12-01') & (dataset['created_at'] < '2019-01-01')]
print("Min:", min(df['created_at']))
print("MAx:", max(df['created_at']))
print(df.shape)

### Tweets por Dataset
df['dataset'].value_counts()

### #Tweets por username
df['user_name'].value_counts()

# Enero 2019

df = dataset[(dataset['created_at'] >= '2019-01-01') & (dataset['created_at'] < '2019-02-01')]
print("Min:", min(df['created_at']))
print("MAx:", max(df['created_at']))
print(df.shape)

### Tweets por Dataset
df['dataset'].value_counts()

### #Tweets por username
df['user_name'].value_counts()

# Febrero 2019

df = dataset[(dataset['created_at'] >= '2019-02-01') & (dataset['created_at'] < '2019-03-01')]
print("Min:", min(df['created_at']))
print("MAx:", max(df['created_at']))
print(df.shape)

### Tweets por Dataset
df['dataset'].value_counts()

### #Tweets por username
df['user_name'].value_counts()

# Marzo 2019

df = dataset[(dataset['created_at'] >= '2019-03-01') & (dataset['created_at'] < '2019-04-01')]
print("Min:", min(df['created_at']))
print("MAx:", max(df['created_at']))
print(df.shape)

### Tweets por Dataset
df['dataset'].value_counts()

### #Tweets por username
df['user_name'].value_counts()

# Abril 2019

df = dataset[(dataset['created_at'] >= '2019-04-01') & (dataset['created_at'] < '2019-05-01')]
print("Min:", min(df['created_at']))
print("MAx:", max(df['created_at']))
print(df.shape)

### Tweets por Dataset
df['dataset'].value_counts()

### #Tweets por username
df['user_name'].value_counts()

# Mayo 2019

df = dataset[(dataset['created_at'] >= '2019-05-01') & (dataset['created_at'] < '2019-06-01')]
print("Min:", min(df['created_at']))
print("MAx:", max(df['created_at']))
print(df.shape)

### Tweets por Dataset
df['dataset'].value_counts()

### #Tweets por username
df['user_name'].value_counts()

# Junio 2019

df = dataset[(dataset['created_at'] >= '2019-06-01') & (dataset['created_at'] < '2019-07-01')]
print("Min:", min(df['created_at']))
print("MAx:", max(df['created_at']))
print(df.shape)

### Tweets por Dataset
df['dataset'].value_counts()

### #Tweets por username
df['user_name'].value_counts()

# Julio 2019

df = dataset[(dataset['created_at'] >= '2019-07-01') & (dataset['created_at'] < '2019-08-01')]
print("Min:", min(df['created_at']))
print("MAx:", max(df['created_at']))
print(df.shape)

### Tweets por Dataset
df['dataset'].value_counts()

### #Tweets por username
df['user_name'].value_counts()

df = dataset[(dataset['created_at'] >= '2019-08-01') & (dataset['created_at'] < '2019-09-01')]
print("Min:", min(df['created_at']))
print("MAx:", max(df['created_at']))
print(df.shape)
